+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+  - [744](https://github.com/thoth-pub/thoth/pull/744) - Support accessibility report file upload
 
 ## [[1.4.0]](https://github.com/thoth-pub/thoth/releases/tag/v1.4.0) - 2026-07-16
 ### Added